implement to version 3 v.1

+++ src/main/java/com/example/kanbanbackend/Service/TaskService.java
@@ -1,34 +1,41 @@
 package com.example.kanbanbackend.Service;
 
 
+import com.example.kanbanbackend.Config.LimitConfig;
+import com.example.kanbanbackend.Config.Permission;
+import com.example.kanbanbackend.DTO.StatusDTO.StatusListAllTaskDTO;
 import com.example.kanbanbackend.DTO.StatusDTO.StatusSelectedDTO;
 import com.example.kanbanbackend.DTO.TaskDTO.TaskAddEditDTO;
 import com.example.kanbanbackend.DTO.TaskDTO.TaskDTO;
 import com.example.kanbanbackend.DTO.TaskDTO.TaskSelectedDTO;
 import com.example.kanbanbackend.DTO.TaskDTO.TaskV3.TaskBoardDTO;
+import com.example.kanbanbackend.DTO.TaskDTO.TaskV3.TaskBoardtestDTO;
+import com.example.kanbanbackend.Entitites.Primary.Board;
+import com.example.kanbanbackend.Entitites.Primary.Config;
 import com.example.kanbanbackend.Entitites.Primary.Status;
 import com.example.kanbanbackend.Entitites.Primary.Task;
 import com.example.kanbanbackend.Exception.BadRequestException;
 import com.example.kanbanbackend.Exception.BadRequestWithFieldException;
 import com.example.kanbanbackend.Exception.ItemNotFoundDelUpdate;
 import com.example.kanbanbackend.Exception.ItemNotFoundException;
+import com.example.kanbanbackend.Repository.Primary.BoardRepository;
+import com.example.kanbanbackend.Repository.Primary.ConfigRepository;
 import com.example.kanbanbackend.Repository.Primary.StatusRepository;
 import com.example.kanbanbackend.Repository.Primary.TaskRepository;
-import com.example.kanbanbackend.Config.LimitConfig;
-import com.example.kanbanbackend.Config.Permission;
 import jakarta.transaction.Transactional;
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 
+import java.sql.Timestamp;
 import java.time.LocalDateTime;
 import java.time.ZoneOffset;
 import java.time.ZonedDateTime;
 import java.time.format.DateTimeFormatter;
 import java.util.ArrayList;
+import java.util.Date;
 import java.util.List;
-import java.util.Locale;
 import java.util.Optional;
 
 @Service
@@ -41,16 +48,22 @@ public class TaskService {
     private TaskRepository repository;
 
     @Autowired
+    private ConfigRepository configRepository;
+
+    @Autowired
     private StatusRepository statusRepository;
+
+    @Autowired
+    private BoardRepository boardRepository;
 
     @Autowired
     private StatusService statusService;
 
     @Autowired
-    private Permission permission;
+    private LimitService limitService;
 
 
-    public List<TaskDTO> getAllTodo(List<String> filterStatuses, String sortBy, String sortDirection){
+    public List<TaskDTO> getAllTodo(List<String> filterStatuses, String sortBy, String sortDirection) {
 
         List<Task> taskList = new ArrayList<>();
         Sort.Direction direction = sortDirection.equalsIgnoreCase("ASC") ? Sort.Direction.ASC : Sort.Direction.DESC;
@@ -58,63 +71,79 @@ public class TaskService {
         if ("statusName".equals(sortBy)) {
             sortBy = "taskStatus.statusName";
         }
-        Sort sort = Sort.by(direction,sortBy);
-        if(filterStatuses != null){
-                taskList.addAll(repository.findAllByStatusNamesSorted(filterStatuses,sort)) ;
-            return  listMapper.mapList(taskList, TaskDTO.class);
+        Sort sort = Sort.by(direction, sortBy);
+        if (filterStatuses != null) {
+            taskList.addAll(repository.findAllByStatusNamesSorted(filterStatuses, sort));
+            return listMapper.mapList(taskList, TaskDTO.class);
         }
         List<Task> tasks = repository.findAll(sort);
-        return  listMapper.mapList(tasks,TaskDTO.class);
+        return listMapper.mapList(tasks, TaskDTO.class);
     }
 
-    public TaskSelectedDTO getTaskById(int id) throws ItemNotFoundException {
-        Task task =  repository.findById(id).orElseThrow(() -> new ItemNotFoundException("Task "+ id +" does not exist !!!" ));
+    public TaskSelectedDTO getTaskById(String boardId, int id) throws ItemNotFoundException {
+//        Task task =  repository.findById(id).orElseThrow(() -> new ItemNotFoundException("Task "+ id +" does not exist !!!" ));
+        Task task = repository.findTaskByBoard_IdAndId(boardId, id);
+        if (task == null) throw new ItemNotFoundException("Task id: " + id + " or boardId: "+ boardId +" doesn't exist !!!");
 
-        LocalDateTime createdDateTime = LocalDateTime.parse(task.getCreatedOn(), DateTimeFormatter.ofPattern("uuuu-MM-dd HH:mm:ss", Locale.ROOT));
-        LocalDateTime updatedDateTime = LocalDateTime.parse(task.getUpdatedOn(), DateTimeFormatter.ofPattern("uuuu-MM-dd HH:mm:ss", Locale.ROOT));
+        // Assuming `getCreatedOn` and `getUpdatedOn` return `Timestamp`
+        Timestamp createdOnTimestamp = task.getCreatedOn();
+        Timestamp updatedOnTimestamp = task.getUpdatedOn();
 
+
+        // Convert Timestamp to LocalDateTime
+        LocalDateTime createdDateTime = createdOnTimestamp.toLocalDateTime();
+        LocalDateTime updatedDateTime = updatedOnTimestamp.toLocalDateTime();
+
+        // Convert LocalDateTime to ZonedDateTime in UTC
         ZonedDateTime createdUTC = createdDateTime.atZone(ZoneOffset.UTC);
         ZonedDateTime updatedUTC = updatedDateTime.atZone(ZoneOffset.UTC);
+        System.out.println(createdUTC);
+
+        // Convert ZonedDateTime to ISO 8601 format string
+        String createdOnISO = createdUTC.format(DateTimeFormatter.ISO_INSTANT);
+        String updatedOnISO = updatedUTC.format(DateTimeFormatter.ISO_INSTANT);
+
+        TaskSelectedDTO dto = mapper.map(task, TaskSelectedDTO.class);
+        dto.setCreatedOn(createdOnISO);
+        dto.setUpdatedOn(updatedOnISO);
 
 
-
-        task.setCreatedOn(DateTimeFormatter.ISO_INSTANT.format(createdUTC));
-        task.setUpdatedOn(DateTimeFormatter.ISO_INSTANT.format(updatedUTC));
-
-        return mapper.map(task, TaskSelectedDTO.class);
+        return dto;
 
 
     }
 
     @Transactional
-    public TaskDTO createTask(TaskAddEditDTO newTaskDTO) {
-        Status statusfind = statusRepository.findById(newTaskDTO.getTaskStatusId()).orElseThrow(() -> new BadRequestException("Status "+ newTaskDTO.getTaskStatusId() +" does not exist !!!" ));
-        if(LimitConfig.isLimit && permission.canEditOrDelete(newTaskDTO.getTaskStatusId())){
-            List<Task> listTasks = repository.findByTaskStatus(statusfind);
-            if(listTasks.size() >= LimitConfig.number) {
-                throw new BadRequestException("The Status has on the limit ("+ LimitConfig.number +")s");
-            }
-        }
-        Task newTask = mapper.map(newTaskDTO,Task.class);
+    public TaskDTO createTask(String boardId,TaskAddEditDTO newTaskDTO) {
+        Status statusfind = statusRepository.findStatusByBoard_IdAndId(boardId,newTaskDTO.getTaskStatusId());
+        if(statusfind == null) throw new BadRequestWithFieldException("status", "Status id "+newTaskDTO.getTaskStatusId()+" or Board id"+ boardId +" does not exist");
+
+        limitService.CheckLimitTask(boardId,statusfind.getId());
+
+
+        newTaskDTO.setBoardId(boardId);
+        Task newTask = mapper.map(newTaskDTO, Task.class);
         newTask.setTaskTitle(newTaskDTO.getTaskTitle() == null ? null : newTaskDTO.getTaskTitle().trim());
-        newTask.setTaskDescription(newTaskDTO.getTaskDescription() == null || newTaskDTO.getTaskDescription().isEmpty()  ? null  : newTaskDTO.getTaskDescription().trim());
+        newTask.setTaskDescription(newTaskDTO.getTaskDescription() == null || newTaskDTO.getTaskDescription().isEmpty() ? null : newTaskDTO.getTaskDescription().trim());
         newTask.setTaskAssignees(newTaskDTO.getTaskAssignees() == null || newTaskDTO.getTaskAssignees().isEmpty() ? null : newTaskDTO.getTaskAssignees().trim());
         repository.saveAndFlush(newTask);
-        Status status = mapper.map(statusfind,Status.class);
+        Status status = mapper.map(statusfind, Status.class);
         newTask.setTaskStatus(status);
         return mapper.map(newTask, TaskDTO.class);
     }
+
     @Transactional
-    public TaskDTO updateTask(Integer taskId, TaskAddEditDTO editedTask ){
-        if(editedTask.getTaskTitle() == null) throw new BadRequestWithFieldException("titie","must not be null");
-        if(LimitConfig.isLimit && permission.canEditOrDelete(editedTask.getTaskStatusId())){
-            List<Task> listTasks = repository.findByTaskStatusId(editedTask.getTaskStatusId());
-            if(listTasks.size() >= LimitConfig.number) {
-                throw new BadRequestException("The Status has on the limit (" + LimitConfig.number + ")s You can't edit !!!");
-            }
-        }
-       Status isExited = statusRepository.findById(editedTask.getTaskStatusId()).orElseThrow(() -> new BadRequestWithFieldException("status","does not exist"));
-        Task oldTask = repository.findById(taskId).orElseThrow(() -> new ItemNotFoundDelUpdate( "NOT FOUND "));
+    public TaskDTO updateTask(String boardId,Integer taskId, TaskAddEditDTO editedTask) {
+        if (editedTask.getTaskTitle() == null) throw new BadRequestWithFieldException("titie", "must not be null");
+
+        limitService.CheckLimitTask(boardId,editedTask.getTaskStatusId());
+
+        Status isExited = statusRepository.findStatusByBoard_IdAndId(boardId,editedTask.getTaskStatusId());
+        if(isExited == null) throw new BadRequestWithFieldException("status", "Status id "+editedTask.getTaskStatusId()+"or Board id"+ boardId +" does not exist");
+
+
+        Task oldTask = repository.findTaskByBoard_IdAndId(boardId, taskId);
+        if (oldTask == null) throw new ItemNotFoundException("Task id: " + taskId + " or boardId: "+ boardId +" doesn't exist !!!");
         Optional.ofNullable(editedTask.getTaskTitle())
                 .map(String::trim)
                 .ifPresent(editedTask::setTaskTitle);
@@ -129,28 +158,36 @@ public class TaskService {
         oldTask.setTaskAssignees(editedTask.getTaskAssignees() != null ? editedTask.getTaskAssignees() : oldTask.getTaskAssignees());
         oldTask.setTaskStatus(editedTask.getTaskStatusId() != null ? isExited : oldTask.getTaskStatus());
         oldTask.setTaskDescription(editedTask.getTaskDescription() != null ? editedTask.getTaskDescription() : oldTask.getTaskDescription());
-        repository.save(oldTask);
-        StatusSelectedDTO newStatus = statusService.getStatusById(oldTask.getTaskStatus().getId());
-        Status status = mapper.map(newStatus,Status.class);
-        oldTask.setTaskStatus(status);
-        return mapper.map(oldTask, TaskDTO.class);
+        Date now = new Date();
+        Timestamp timestamp = new Timestamp(now.getTime());
+        oldTask.setUpdatedOn(timestamp);
+
+        repository.saveAndFlush(oldTask);
+        StatusSelectedDTO newStatus = statusService.getStatusByIdAndBoardId(boardId,oldTask.getTaskStatus().getId());
+        TaskBoardtestDTO dto = mapper.map(oldTask,TaskBoardtestDTO.class);
+        dto.setTaskStatus(newStatus);
+
+        return mapper.map(dto, TaskDTO.class);
     }
 
     @Transactional
-    public TaskDTO deleteTask(Integer delId){
-        Task delTask = repository.findById(delId).orElseThrow(() -> new ItemNotFoundDelUpdate( "NOT FOUND"));
+    public TaskDTO deleteTask(String boardId,Integer delId) {
+        Task delTask = repository.findTaskByBoard_IdAndId(boardId, delId);
+        if (delTask == null) throw new ItemNotFoundDelUpdate("NOT FOUND");
+
+
         repository.delete(delTask);
         return mapper.map(delTask, TaskDTO.class);
     }
 
     // Version 3
 
-    public List<TaskBoardDTO> getTaskofBoard(Integer boardId){
-            List<Task> taskBoard = repository.findTaskByBoard_Id(boardId);
-            if(taskBoard.isEmpty()){
-                throw new BadRequestException("This board has no task !!");
-            }
-            return listMapper.mapList(taskBoard,TaskBoardDTO.class );
+    public List<TaskBoardDTO> getTaskofBoard(String boardId) {
+        List<Task> taskBoard = repository.findTaskByBoard_Id(boardId);
+        if (taskBoard.isEmpty()) {
+            throw new BadRequestException("This board has no task !!");
+        }
+        return listMapper.mapList(taskBoard, TaskBoardDTO.class);
     }
 
 }
